Pass commit hash to benchmarker
To make sure the selected PR is actually properly built, the
headless-bench script now takes a commit hash as a parameter.

Now instead of having to ship a version of alacritty, the benchmarking
script automatically clones the repository and checks out the desired
commit, then runs the benchmarks with that.