Run benchmarks only on commits/PRs to master
Instead of always running the benchmarks, it is now only run when a
PR/commit is sent with the `master` branch as a target.

The resulting directory structure has also been changed, PRs are now
stored in `./results/pr-{NUM}/{TIME}-{COMMIT}` and the results for
commits to the master branch are stored it
`./results/master/{TIME}-{COMMIT}`.