Add padding area around small charts
If charts do not have a certain height to it, but the top/bottom value
is less than 10% of the average of all values, additional room will be
added to the chart to make sure minor fluctuations do not look like
massive regressions.

This also adds a custom chart style.